+++ src/sempipe/cli/root.py
@@ -37,6 +37,7 @@ from sempipe.cli.run_cmd import run_command
 from sempipe.cli.sample_cmd import sample_command
 from sempipe.cli.schema_cmd import schema_command
 from sempipe.cli.screens import WELCOME
+from sempipe.cli.sort_cmd import sort_command
 from sempipe.cli.split_cmd import split_command
 from sempipe.cli.summarize_cmd import summarize_command
 from sempipe.cli.top_k_cmd import top_k_command
@@ -91,6 +92,7 @@ cli.add_command(where_command)
 cli.add_command(summarize_command)
 cli.add_command(sample_command)
 cli.add_command(getschema_command)
+cli.add_command(sort_command)
 cli.add_command(auth_command)
 cli.add_command(cite_command)
 cli.add_command(echo_command)